Firewall the remote CEF port by default
By default, once CEF is enabled, in addition to the `localhost:8080` CEF
is also available on `0.0.0.0:8081`. As this is potentially insecure,
as everyone on the network would be able to interact with a user's deck,
PluginLoader now disables access to the port 8081 from anything but
`localhost` via iptables by default.
For developers, the port can be re-enabled by placing an empty file with
the name `allow_cef_from_network` in the same directory as the
`PLUGIN_PATH` folder (usually `~/homebrew`).
PluginLoader will pick up on this and skip firewalling off the CEF port.