feat: add "open" action to plugins settings
This allows opening plugins even if they are hidden in the quick access
list.

This solves the use case where plugins are hidden from the quick access
list because access to their panels is not used very often. For example,
plugin "ProtonDB Badges" is fire-configure-once-and-forget, so it makes
sense to hide it so it doesn't waste space and attention.

But this doesn't mean the user will never want to access hidden plugin
panels to e.g. briefly change configuration. Currently the only way to
access the plugin content panel is to unhide the plugin, access it in
the quick access list, and then hide it again.

Resolves #734